:recycle: refactor types
