+++ src/whois/mod.rs
@@ -1,13 +1,15 @@
 mod args;
 
 pub(crate) use args::Args;
+use inkanim::{
+    anim::InkAnimAnimationLibraryResource,
+    widget::{inkWidgetLibraryResource, WidgetTree},
+};
 use term_table::{
     row::Row,
     table_cell::{Alignment, TableCell},
     Table, TableStyle,
 };
-
-use inkanim::{inkWidgetLibraryResource, InkAnimAnimationLibraryResource, WidgetTree};
 
 pub(crate) fn whois(
     args: Args,
